+++ crates/iota-config/Cargo.toml
@@ -30,7 +30,6 @@ tracing.workspace = true
 # internal dependencies
 consensus-config.workspace = true
 iota-genesis-common.workspace = true
-iota-grpc-api.workspace = true
 iota-keys.workspace = true
 iota-names.workspace = true
 iota-rest-api.workspace = true
