+++ crates/iota-grpc-api/Cargo.toml
@@ -25,6 +25,7 @@ tonic.workspace = true
 tracing.workspace = true
 
 # internal dependencies
+iota-config.workspace = true
 iota-grpc-types.workspace = true
 iota-json-rpc-types.workspace = true
 iota-types.workspace = true
