+++ crates/iota-swarm-config/Cargo.toml
@@ -26,7 +26,6 @@ tracing.workspace = true
 # internal dependencies
 iota-config.workspace = true
 iota-genesis-builder.workspace = true
-iota-grpc-api.workspace = true
 iota-macros.workspace = true
 iota-names.workspace = true
 iota-protocol-config.workspace = true
